Rename quantaq_py to quantaq_tools
