+++ quantaq_tools/toolkit/concat.py
@@ -1,7 +1,7 @@
 from loguru import logger
 import pandas as pd
 
-from quantaq_py.utilities import safe_load, fix_timestamps
+from quantaq_tools.utilities import safe_load, fix_timestamps
 
 def concat_files(files):
     """Concatenate multiple files into a single DataFrame and sort based on a
